[FIX] TagDetection::getRelativeTransform() double<->float missmatch fixed.
HINT: causes an assertion in cv::Rodrigues() and the returned transformation T had  wrong values!

+++ apriltags/src/TagDetection.cc
@@ -98,14 +98,14 @@ Eigen::Matrix4d TagDetection::getRelativeTransform(double tag_size, double fx, d
                            0,  0,  1);
   cv::Vec4f distParam(0,0,0,0); // all 0?
   cv::solvePnP(objPts, imgPts, cameraMatrix, distParam, rvec, tvec);
-  cv::Matx33d r;
+  cv::Matx33f r;
   cv::Rodrigues(rvec, r);
   Eigen::Matrix3d wRo;
   wRo << r(0,0), r(0,1), r(0,2), r(1,0), r(1,1), r(1,2), r(2,0), r(2,1), r(2,2);
 
   Eigen::Matrix4d T; 
   T.topLeftCorner(3,3) = wRo;
-  T.col(3).head(3) << tvec.at<double>(0), tvec.at<double>(1), tvec.at<double>(2);
+  T.col(3).head(3) << tvec.at<float>(0), tvec.at<float>(1), tvec.at<float>(2);
   T.row(3) << 0,0,0,1;
 
   return T;
